requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
# Analysis report: DB, queue, LLM
sqlalchemy>=2.0
psycopg2-binary
celery[redis]
redis
openai
httpx
# Generation: Azure Speech, S3
boto3
azure-cognitiveservices-speech
pyyaml
pytest

